+++ GHEtool/test/test_examples.py
@@ -103,3 +103,18 @@ def test_turbocollector(monkeypatch):
     monkeypatch.setattr(plt, 'show', lambda: None)
     from GHEtool.Examples.turbocollector import create_graphs
     create_graphs()
+
+
+def test_optimal_borehole_config():
+    from GHEtool.Examples.optimal_borehole_configuration import borefield_case_1, borefield_case_2, \
+        borefield_case_3, borefield_case_4, borefield_office, borefield_auditorium, borefield_swimming_pool, \
+        borefield_case_1_flow_rate
+
+    borefield_case_1()
+    borefield_case_2()
+    borefield_case_3()
+    borefield_case_4()
+    borefield_office()
+    borefield_auditorium()
+    borefield_swimming_pool()
+    borefield_case_1_flow_rate()